Fixed many bugs!

+++ shoppingcart/Clothing.java
@@ -22,7 +22,7 @@ public class Clothing extends PurchaseItem
 	 * @param itemQuantity Item quantity.
 	 * @param itemWeight Item weight.
 	 */
-	public Clothing(String itemName, double itemPrice, int itemQuantity, int itemWeight) {
+	public Clothing(String itemName, double itemPrice, int itemQuantity, double itemWeight) {
 		super(itemName, itemPrice, itemQuantity, itemWeight);
 		//premium shipping is not available
 		isPremium = false;
